% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Calculate the estimated error of the estimate}
\usage{
accuracy(P, hatP, T, thr = NULL)
}
\arguments{
\item{P}{input set}

\item{hatP}{estiamtes of set P}

\item{T}{a positive number to calculate thr}

\item{thr}{if \eqn{thr=NULL}, let \eqn{thr=log(T)}}
}
\value{
the estimated error of the estimate
}
\description{
Calculate the estimated error of the estimate
}
