% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSCOV.R
\name{BSCOV}
\alias{BSCOV}
\title{Using Wild Sparsified Binary Segmentation to find chang points in the second order structure in both factors and idiosyncratic errors.}
\usage{
BSCOV(
  x,
  normalisation = FALSE,
  r = NULL,
  bn.op = 2,
  sig.lev = 1,
  dw = NULL,
  rule = NULL,
  SN.op = 0,
  refact.op = 0,
  common.norm = 2,
  idio.norm = 2,
  common.norm.thr = 0,
  idio.norm.thr = Inf,
  norm.op = FALSE,
  SSIC.slack = c(1/2, 1/2),
  max.q = NULL,
  WBS = 1,
  SBS = 1,
  M = 500,
  input.op = 1,
  idio.diag = FALSE,
  do.parallel = TRUE,
  no.proc = 2,
  setseed = 0
)
}
\arguments{
\item{x}{input time series matrix, with each row representing a time series}

\item{normalisation}{if \eqn{normalisation=TRUE}, normalized the input data}

\item{r}{the number of factors, if \eqn{r=NULL}, screening over a range of factor number candidates is performed as described in the paper}

\item{bn.op}{an index number for the information criterion-based estimator of Bai and Ng (2002) for the number of factors}

\item{sig.lev}{not used now}

\item{dw}{trims off the interval of consideration in the binary segmentation algorithm and determines the minimum length of a stationary segment; if dw=NULL, a default value is chosen as described in the Appendix of Barigozzi, Cho & Fryzlewicz (2016)}

\item{rule}{the depth of a binary tree for change-point analysis, see the Appendix of Barigozzi, Cho & Fryzlewicz (2016)}

\item{SN.op}{if \eqn{SN.op==1} Self Normalized; \eqn{SN.op==2} mean of absolute; \eqn{SN.op==3} \eqn{mad(abs(diff(x)-mad(diff(x))))} ...}

\item{refact.op}{if \eqn{refact.op=TRUE}, redo the PCA on each segment after the breaks in the factors are found}

\item{common.norm}{norm used in the  CUSUM statistics aggregation of factors}

\item{idio.norm}{norm used in the  CUSUM statistics aggregation of idiosyncratic errors}

\item{common.norm.thr}{threshold used in the CUSUM statistics of factors}

\item{idio.norm.thr}{threshold used in the CUSUM statistics of idiosyncratic errors}

\item{norm.op}{if \eqn{idio.diag=TRUE} the threshold is multiplied by \eqn{((e-s+1)/T)^0.5}}

\item{SSIC.slack}{positive constant used in the SSIC}

\item{max.q}{the maximum number of factors, if \eqn{max.q=NULL}, a default value is chosen as described in the paper}

\item{WBS}{Wild Binary Segmentation}

\item{SBS}{Sparsified Binary Segmentation}

\item{M}{the number of random intervals}

\item{input.op}{options of data tramsformation. if \eqn{input.op=1},  second moments, if \eqn{input.op=2}, square of difference and square of sum, if \eqn{input.op=3}, wavelet transformation}

\item{idio.diag}{if \eqn{idio.diag=TRUE}, only the diagonal wavelet-transform is employed in order to generate the panel of statistics from the idiosyncratic components}

\item{do.parallel}{if \eqn{do.parallel=TRUE}, a set of copies of R running in parallel are created and used for bootstrap procedure}

\item{no.proc}{sets the number of processes to be spawned when do.parallel=TRUE}

\item{setseed}{sets the ramdon seed}
}
\value{
change points
\itemize{
 \item{"gfm"}{information of factor model}
 \item{"para"}{Parameters used in the algotrithm}
 \item{"hat.vep"}{estimate of idiosyncratic components}
 \item{"common.seg.res"}{information of common component breaks}
 \item{"idio.seg.res"}{information of idiosyncratic component breaks}
 \item{"common.est.cps"}{location of common component breaks}
 \item{"idio.est.cps"}{location idiosyncratic component breaks}
}
}
\description{
Using Wild Sparsified Binary Segmentation to find chang points in the second order structure in both factors and idiosyncratic errors.
}
\author{
Yu-Ning Li \email{yl3409@york.ac.uk}
}
