% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul.cusum.R
\name{calcul.cusum}
\alias{calcul.cusum}
\title{Construction of CUSUM stiatistics}
\usage{
calcul.cusum(x, dw = 0, alpha = 0.5, SN.op = 0)
}
\arguments{
\item{x}{input time series matrix, with each row representing a time series}

\item{dw}{trims off the interval of consideration in the binary segmentation algorithm and determines the minimum length of a stationary segment; if dw=NULL, a default value is chosen as described in the Appendix of Barigozzi, Cho & Fryzlewicz (2016)}

\item{alpha}{CUSUM parameter}

\item{SN.op}{if \eqn{SN.op==1} Self Normalized; \eqn{SN.op==2} mean of absolute; \eqn{SN.op==3} \eqn{mad(abs(diff(x)-mad(diff(x))))} ...}
}
\value{
CUSUM stiatistics
}
\description{
Construction of CUSUM stiatistics
}
