% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul.input.R
\name{calcul.input}
\alias{calcul.input}
\title{Data transformataion before construction of CUSUM stiatistics}
\usage{
calcul.input(raw, idio.diag = FALSE, input.op = 1)
}
\arguments{
\item{raw}{input time series matrix, with each row representing a time series}

\item{idio.diag}{if(\eqn{idio.diag=TRUE}) only diagnal elements (variance) are used}

\item{input.op}{options of data tramsformation. if \eqn{input.op=1},  second moments, if \eqn{input.op=2}, square of difference and square of sum, if \eqn{input.op=3}, wavelet transformation}
}
\value{
the transformed data
}
\description{
Data transformataion before construction of CUSUM stiatistics
}
