% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul.matrnd.R
\name{calcul.matrnd}
\alias{calcul.matrnd}
\title{Using Wild Sparsified Binary Segmentation to find chang points in the second order structure in both factors and idiosyncratic errors.}
\usage{
calcul.matrnd(y, dw, screen.id, matrnd, normf, SN.op, do.parallel)
}
\arguments{
\item{y}{input time series matrix, with each row representing a time series}

\item{dw}{trims off the interval of consideration in the binary segmentation algorithm and determines the minimum length of a stationary segment; if dw=NULL, a default value is chosen as described in the Appendix of Barigozzi, Cho & Fryzlewicz (2016)}

\item{screen.id}{id given by screening}

\item{SN.op}{if \eqn{SN.op==1} Self Normalized; \eqn{SN.op==2} mean of absolute; \eqn{SN.op==3} \eqn{mad(abs(diff(x)-mad(diff(x))))} ...}

\item{do.parallel}{if \eqn{do.parallel=TRUE}, a set of copies of R running in parallel are created and used for bootstrap procedure}

\item{mathrnd}{random intervals}
}
\value{
matrix with first row MaxCUSUM, second row location of MaxCUSUM
}
\description{
Using Wild Sparsified Binary Segmentation to find chang points in the second order structure in both factors and idiosyncratic errors.
}
\author{
Yu-Ning Li \email{yl3409@york.ac.uk}
}
