% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hausdorff.R
\name{Hausdorff}
\alias{Hausdorff}
\title{Calculate the Hausdorff distance
'}
\usage{
Hausdorff(P, Q, MV = Inf)
}
\arguments{
\item{P}{input set}

\item{Q}{infput set}

\item{MV}{the distance defined for empty set}
}
\value{
Hausdorff distance
}
\description{
Calculate the Hausdorff distance
'
}
