% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.thr.R
\name{select.thr}
\alias{select.thr}
\title{Select the thresholding parameter using the  quantile of CUSUM stiatistics}
\usage{
select.thr(x, dw = 1, SN.op = 0, M = 1, select = seq(0.99, 1, 0.001))
}
\arguments{
\item{x}{input time series matrix, with each row representing a time series}

\item{dw}{trims off the interval of consideration in the binary segmentation algorithm and determines the minimum length of a stationary segment; if dw=NULL, a default value is chosen as described in the Appendix of Barigozzi, Cho & Fryzlewicz (2016)}

\item{SN.op}{if \eqn{SN.op==1} Self Normalized; \eqn{SN.op==2} mean of absolute; \eqn{SN.op==3} \eqn{mad(abs(diff(x)-mad(diff(x))))} ...}

\item{M}{number of re-sampling}

\item{select}{quantile position}
}
\value{
quantile of CUSUM stiatistics
}
\description{
Select the thresholding parameter using the  quantile of CUSUM stiatistics
}
