% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSIC.R
\name{SSIC}
\alias{SSIC}
\title{Using SSIC selects the number of breaks}
\usage{
SSIC(x, mat, SSIC.slack = c(1/2, 1/2), SIC.const = NULL)
}
\arguments{
\item{x}{input time series matrix, with each row representing a time series}

\item{mat}{information of change points}

\item{SSIC.slack}{two positive constants as the thresholding parameters}

\item{SIC.const}{pennalty function}
}
\value{
information of breaks with one of the thresholding parameters \itemize{
 \item{"pen"}{pennalty function}
 \item{"ic.curve"}{information criteria}
 \item{"dim.cpt.ic"}{dimension that has the breaks}
 \item{"no.cpt.ic "}{number of breaks}
 \item{"cpt.ic"}{location of breaks}
}
}
\description{
Using SSIC selects the number of breaks
}
